backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
supabase
google-generativeai
langgraph
langchain-community
python-dotenv
httpx
